rename: 버전 구분으로 패키지 변경
