+++ src/main/java/com/spring/board/springboard/common/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.spring.board.springboard.domain;
+package com.spring.board.springboard.common;
 
 public record ErrorResponse(String errorMessage) {
 }
