+++ src/main/java/com/spring/board/springboard/common/ValidationErrorResponse.java
@@ -1,4 +1,4 @@
-package com.spring.board.springboard.domain;
+package com.spring.board.springboard.common;
 
 public record ValidationErrorResponse (String field, String message) {
