+++ src/main/java/com/spring/board/springboard/user/controller/authenticate/session/Session.java
@@ -1,4 +1,4 @@
-package com.spring.board.springboard.user.controller.authenticate;
+package com.spring.board.springboard.user.controller.authenticate.session;
 
 import com.spring.board.springboard.user.exception.SessionCreateException;
