+++ src/test/java/com/spring/board/springboard/post/controller/PostControllerV2Test.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.spring.board.springboard.post.domain.dto.PostCreateRequestDto;
 import com.spring.board.springboard.post.domain.dto.PostDetailResponseDto;
 import com.spring.board.springboard.post.service.PostService;
-import com.spring.board.springboard.user.controller.authenticate.Session;
-import com.spring.board.springboard.user.controller.authenticate.SessionManager;
+import com.spring.board.springboard.user.controller.authenticate.session.Session;
+import com.spring.board.springboard.user.controller.authenticate.session.SessionManager;
 import com.spring.board.springboard.user.domain.Hobby;
 import com.spring.board.springboard.user.domain.Member;
 import com.spring.board.springboard.user.domain.dto.MemberDetailResponseDto;
@@ -24,7 +24,7 @@ import org.springframework.http.MediaType;
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.test.web.servlet.MockMvc;
 
-import static com.spring.board.springboard.user.controller.authenticate.CookieUtils.createUserInfoCookie;
+import static com.spring.board.springboard.user.controller.authenticate.cookie.CookieUtils.createUserInfoCookie;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
